refactor(metal): merge MetalF16Backend into MetalBackend via tagged buffers
Previous design added a parallel `MetalF16Backend` type to carry fp16
weight support. That conflated two orthogonal axes: hardware (Metal /
CUDA / CPU) and precision (f32 / f16 / future INT8 / bf16). Every new
dtype would have forced a new backend type, which is how the existing
GPTQ handling already demonstrated *not* to do it — GPTQ lives inside
`Backend` via `type GptqStore` + `load_gptq` / `gemm_gptq`, one entry
per hardware.

Unifies the two:
- `MetalBackend::Buffer` becomes `MetalBuf { raw, dtype, n }`
- `Dtype` enum is the extension point; INT8 / bf16 only need a variant
  + shader wiring, not a new backend
- `gemm` and `embedding_lookup` dispatch on buffer dtype
- `from_slice` always returns F32 (activations)
- `from_weight_bytes` consults `FERRUM_METAL_DTYPE` + a size threshold:
  tiny tensors (norm weights, biases) stay F32 so f32-only shaders keep
  working; big tensors (> 1 M elems) promote to F16 when requested

Engine / loader / tests switch back to the single `MetalBackend`; the
env var stays but is now interpreted inside the backend rather than
toggling the concrete type.

Net diff: -626 / +401 lines of backend code, minus the duplicated
MetalF16Backend file.

Parity tests pass:
- CPU ↔ Metal (default f32): bit-exact (6 tokens)
- CPU ↔ Metal (FERRUM_METAL_DTYPE=f16): bit-exact (6 tokens)
- Size-threshold probe: tiny→F32, big→F16
