+++ crates/ferrum-kernels/src/backend/metal.rs
@@ -1,13 +1,35 @@
-//! Metal backend — pipelined command buffer with manual retain/release.
+//! Metal backend — one unified backend with dtype-tagged buffers.
 //!
-//! Pattern from llama.cpp: [queue commandBuffer] + [cmd_buf retain].
-//! One command buffer per forward pass; GPU ops encode into it; sync() commits.
+//! Precision is carried as a runtime tag on [`MetalBuf`] rather than a
+//! separate backend type. Ops branch on buffer dtype where it matters
+//! (`gemm`, `embedding_lookup`); the rest assert F32 and go through the
+//! existing f32 shaders.
+//!
+//! The default policy:
+//!   - Activations (`alloc`, `from_slice`, scratch) are F32 — precision-
+//!     sensitive paths (norm / rope / attention) stay unchanged.
+//!   - Weights loaded via `from_weight_bytes` go F16 for tensors past a
+//!     size threshold when `FERRUM_METAL_DTYPE=f16` is set; otherwise
+//!     they stay F32.
+//!
+//! The same shape generalises: an INT8 / GPTQ variant would add another
+//! `Dtype` case + the corresponding shader, without introducing a new
+//! `Backend` type. Hardware (Metal / CUDA / CPU) and precision remain
+//! orthogonal, matching how the `Backend::GptqStore` + `gemm_gptq`
+//! trait plumbing already handles INT4 on CUDA.
+//!
+//! Command buffer lifecycle mirrors llama.cpp: one retained buffer per
+//! forward pass that accumulates encodes; `sync()` commits and waits.
 
-use super::{AttnConfig, Backend};
+use super::{AttnConfig, Backend, SrcDtype};
 use ferrum_attention::metal::pipelines::MetalPipelines;
 use ferrum_attention::AttentionParams;
-use metal::Device;
+use half::{bf16, f16};
+use metal::{Device, MTLResourceOptions};
+use std::ffi::c_void;
 use std::sync::OnceLock;
+
+// ── Shared Metal state ────────────────────────────────────────────────
 
 struct MetalState {
     pipes: MetalPipelines,
@@ -19,26 +41,101 @@ fn st() -> &'static MetalState {
     })
 }
 
-pub struct MetalBackend;
+// ── Dtype tag + tagged buffer ─────────────────────────────────────────
 
-/// Metal context — accumulates GPU work in a single command buffer across
-/// multiple Backend method calls. `sync()` commits and creates a fresh one on demand.
+/// Element storage type for a [`MetalBuf`]. Same shape generalises to INT8
+/// / bf16 etc. when their shaders land — just add a variant + wire it in
+/// the op dispatches.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Dtype {
+    F32,
+    F16,
+}
+
+impl Dtype {
+    pub const fn bytes_per_elem(self) -> usize {
+        match self {
+            Dtype::F32 => 4,
+            Dtype::F16 => 2,
+        }
+    }
+}
+
+/// Metal device buffer with a runtime dtype tag and logical element count.
 ///
-/// The `&'static CommandBufferRef` lifetime is safe because the command queue lives in
-/// a `OnceLock` (see [`st`]) for the program's lifetime; autoreleased command buffers
-/// are retained for the duration of the queue.
+/// Two buffers of identical raw bytes but different `dtype` are treated as
+/// different types for shader selection. `n` is the number of logical
+/// elements; `raw.length()` is `n * dtype.bytes_per_elem()`.
+pub struct MetalBuf {
+    raw: metal::Buffer,
+    dtype: Dtype,
+    n: usize,
+}
+
+impl MetalBuf {
+    pub fn raw(&self) -> &metal::Buffer {
+        &self.raw
+    }
+    pub fn raw_mut(&mut self) -> &mut metal::Buffer {
+        &mut self.raw
+    }
+    pub fn dtype(&self) -> Dtype {
+        self.dtype
+    }
+    pub fn len(&self) -> usize {
+        self.n
+    }
+    pub fn is_empty(&self) -> bool {
+        self.n == 0
+    }
+    pub fn is_f16(&self) -> bool {
+        matches!(self.dtype, Dtype::F16)
+    }
+
+    #[inline]
+    fn expect_f32<'a>(&'a self, what: &str) -> &'a metal::Buffer {
+        debug_assert!(
+            matches!(self.dtype, Dtype::F32),
+            "{what}: expected F32 buffer, got {:?}",
+            self.dtype
+        );
+        &self.raw
+    }
+    #[inline]
+    fn expect_f32_mut<'a>(&'a mut self, what: &str) -> &'a mut metal::Buffer {
+        debug_assert!(
+            matches!(self.dtype, Dtype::F32),
+            "{what}: expected F32 buffer, got {:?}",
+            self.dtype
+        );
+        &mut self.raw
+    }
+}
+
+// Safety: metal::Buffer wraps a retained NSObject pointer that's safe to move
+// and share across threads (Metal serialises on its command queue).
+unsafe impl Send for MetalBuf {}
+unsafe impl Sync for MetalBuf {}
+
+// ── Context ───────────────────────────────────────────────────────────
+
+/// Metal context — one in-flight command buffer that accumulates encodes
+/// across multiple `Backend` method calls. `sync()` commits + waits and
+/// drops the handle so the next call creates a fresh buffer.
 pub struct MetalContext {
     cmd: Option<&'static metal::CommandBufferRef>,
 }
 
 impl MetalContext {
-    /// Return the current in-flight command buffer, creating one on first use.
+    /// Return the current in-flight command buffer, creating one on first
+    /// use. The `'static` is safe: the command queue lives in a `OnceLock`
+    /// for the program's lifetime, and autoreleased command buffers are
+    /// retained for as long as that queue lives.
     fn cmd(&mut self) -> &'static metal::CommandBufferRef {
         match self.cmd {
             Some(c) => c,
             None => {
                 let c = st().pipes.queue.new_command_buffer();
-                // Erase the queue's lifetime; the queue is static (OnceLock) so this is safe.
                 let c_static: &'static metal::CommandBufferRef =
                     unsafe { std::mem::transmute::<&metal::CommandBufferRef, _>(c) };
                 self.cmd = Some(c_static);
@@ -47,7 +144,6 @@ impl MetalContext {
         }
     }
 
-    /// Commit pending work and wait for completion.
     fn flush(&mut self) {
         if let Some(cmd) = self.cmd.take() {
             cmd.commit();
@@ -56,19 +152,65 @@ impl MetalContext {
     }
 }
 
-/// One-shot helper: open a cmd buffer, encode, commit+wait.
-/// Used only when ctx-driven batching is not needed (tests, isolated ops).
-fn run(f: impl FnOnce(&metal::CommandBufferRef)) {
-    let cmd = st().pipes.queue.new_command_buffer();
-    f(cmd);
-    cmd.commit();
-    cmd.wait_until_completed();
+// ── Policy: should big weights land as f16? ───────────────────────────
+// Cached on first read so each tensor load doesn't re-parse env.
+
+fn prefer_f16_weights() -> bool {
+    static FLAG: OnceLock<bool> = OnceLock::new();
+    *FLAG.get_or_init(|| {
+        std::env::var("FERRUM_METAL_DTYPE")
+            .map(|v| v.eq_ignore_ascii_case("f16"))
+            .unwrap_or(false)
+    })
 }
 
+/// Element-count threshold above which a weight tensor goes to F16 storage
+/// (when `FERRUM_METAL_DTYPE=f16`). Tiny tensors (norm weights, biases) stay
+/// F32 so the existing f32 shaders continue to see f32 inputs.
+///
+/// 1M elements = 4 MB as f32 / 2 MB as f16. Anything smaller saves little
+/// memory and would force the f32-only shaders to sprout f16 variants.
+const F16_MIN_ELEMS: usize = 1_048_576;
+
+// ── Buffer allocation helpers (context-free) ─────────────────────────
+
+fn alloc_f32_raw(n: usize) -> metal::Buffer {
+    st().pipes.buffer_empty(n)
+}
+
+fn buffer_from_f32_slice(data: &[f32]) -> metal::Buffer {
+    st().pipes.buffer_from_data(data)
+}
+
+fn buffer_from_f16_bytes(bytes: &[u8]) -> metal::Buffer {
+    debug_assert_eq!(bytes.len() % 2, 0);
+    st().pipes.device.new_buffer_with_data(
+        bytes.as_ptr() as *const c_void,
+        bytes.len() as u64,
+        MTLResourceOptions::StorageModeShared,
+    )
+}
+
+/// Pack a slice of f32 into a fresh f16 metal::Buffer.
+fn buffer_f16_from_f32(data: &[f32]) -> metal::Buffer {
+    let n = data.len();
+    let mut f16_bytes = vec![0u8; n * 2];
+    for i in 0..n {
+        let h = f16::from_f32(data[i]).to_le_bytes();
+        f16_bytes[i * 2] = h[0];
+        f16_bytes[i * 2 + 1] = h[1];
+    }
+    buffer_from_f16_bytes(&f16_bytes)
+}
+
+// ── Backend impl ──────────────────────────────────────────────────────
+
+pub struct MetalBackend;
+
 impl Backend for MetalBackend {
-    type Buffer = metal::Buffer;
+    type Buffer = MetalBuf;
     type Context = MetalContext;
-    type GptqStore = (); // Metal GPTQ not yet wired; load_gptq/gemm_gptq return unsupported
+    type GptqStore = (); // Metal GPTQ not yet wired; load_gptq/gemm_gptq return unsupported.
 
     fn new_context() -> Self::Context {
         MetalContext { cmd: None }
@@ -76,6 +218,8 @@ impl Backend for MetalBackend {
     fn sync(ctx: &mut Self::Context) {
         ctx.flush();
     }
+
+    // ── GEMM — dispatches on B-weight dtype ──────────────────────────
 
     fn gemm(
         ctx: &mut Self::Context,
@@ -86,21 +230,36 @@ impl Backend for MetalBackend {
         n: usize,
         k: usize,
     ) {
+        let a_buf = a.expect_f32("gemm a");
+        let out_buf = out.expect_f32_mut("gemm out");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
-        if m == 1 {
-            // GEMV: one threadgroup per output col, K-reduction via simd_sum.
-            // Great occupancy for lm_head (N = vocab = 152k for Qwen3).
-            st().pipes.gemv_enc(enc, a, b, out, n, k);
-        } else {
-            // GPU GEMM (simdgroup matrix, 64x32 tiles). Replaces earlier
-            // `cblas_sgemm` fallback: on M1 with 4B-param models the CPU
-            // GEMM path competed with decode threads and thrashed shared-
-            // memory bandwidth. gemm_f32_v2 runs fully on the GPU.
-            st().pipes.gemm_v2(enc, a, b, out, m, n, k);
+        match b.dtype {
+            Dtype::F16 => {
+                // f16 weights — route through the f32-activation / f16-weight kernels.
+                if m == 1 {
+                    st().pipes.gemv_enc_f16w(enc, a_buf, &b.raw, out_buf, n, k);
+                } else {
+                    st().pipes
+                        .gemm_v2_f16w(enc, a_buf, &b.raw, out_buf, m, n, k);
+                }
+            }
+            Dtype::F32 => {
+                if m == 1 {
+                    // GEMV with K-reduction via simd_sum — good for lm_head
+                    // (N = vocab = 152k for Qwen3).
+                    st().pipes.gemv_enc(enc, a_buf, &b.raw, out_buf, n, k);
+                } else {
+                    // GPU simdgroup-matrix tiled GEMM. Replaces an earlier
+                    // cblas_sgemm fallback that serialised against GPU work.
+                    st().pipes.gemm_v2(enc, a_buf, &b.raw, out_buf, m, n, k);
+                }
+            }
         }
         enc.end_encoding();
     }
+
+    // ── Norm / attention / fused ops — all f32 ───────────────────────
 
     fn rms_norm(
         ctx: &mut Self::Context,
@@ -111,6 +270,9 @@ impl Backend for MetalBackend {
         tokens: usize,
         dim: usize,
     ) {
+        let x = x.expect_f32("rms_norm x");
+        let w = w.expect_f32("rms_norm w");
+        let out = out.expect_f32_mut("rms_norm out");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.rms_norm_enc(enc, x, w, out, tokens, dim, eps);
@@ -127,6 +289,10 @@ impl Backend for MetalBackend {
         tokens: usize,
         dim: usize,
     ) {
+        let residual = residual.expect_f32_mut("fused_add_rms_norm residual");
+        let x = x.expect_f32("fused_add_rms_norm x");
+        let w = w.expect_f32("fused_add_rms_norm w");
+        let out = out.expect_f32_mut("fused_add_rms_norm out");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.fused_residual_norm_enc(
@@ -158,6 +324,10 @@ impl Backend for MetalBackend {
             pos_offset,
             sliding_window: cfg.sliding_window,
         };
+        let q = q.expect_f32("flash_attention q");
+        let k = k.expect_f32("flash_attention k");
+        let v = v.expect_f32("flash_attention v");
+        let out = out.expect_f32_mut("flash_attention out");
         let cmd = ctx.cmd();
         st().pipes
             .flash_attn_v2(cmd, q, k, v, out, &p, cfg.kv_seq_stride);
@@ -173,6 +343,8 @@ impl Backend for MetalBackend {
     ) {
         // Blit encoder stays in the same command buffer as neighbouring
         // compute encoders, keeping the single-command-buffer invariant.
+        let src = src.expect_f32("copy_slice src");
+        let dst = dst.expect_f32_mut("copy_slice dst");
         let cmd = ctx.cmd();
         let blit = cmd.new_blit_command_encoder();
         blit.copy_from_buffer(
@@ -192,26 +364,38 @@ impl Backend for MetalBackend {
         out: &mut Self::Buffer,
         dim: usize,
     ) {
-        // CPU table lookup writing to shared buffer; must flush any readers of `out`.
+        // CPU-side gather into the f32 activation output. `ids.len()` is
+        // small (one per batch item); flush any pending GPU work first.
+        let out = out.expect_f32_mut("embedding_lookup out");
         ctx.flush();
         unsafe {
-            let t = std::slice::from_raw_parts(
-                table.contents() as *const f32,
-                table.length() as usize / 4,
-            );
             let o = std::slice::from_raw_parts_mut(out.contents() as *mut f32, ids.len() * dim);
-            for (i, &id) in ids.iter().enumerate() {
-                let s = id as usize * dim;
-                o[i * dim..(i + 1) * dim].copy_from_slice(&t[s..s + dim]);
+            match table.dtype {
+                Dtype::F32 => {
+                    let t = std::slice::from_raw_parts(
+                        table.raw.contents() as *const f32,
+                        table.raw.length() as usize / 4,
+                    );
+                    for (i, &id) in ids.iter().enumerate() {
+                        let s = id as usize * dim;
+                        o[i * dim..(i + 1) * dim].copy_from_slice(&t[s..s + dim]);
+                    }
+                }
+                Dtype::F16 => {
+                    let t = std::slice::from_raw_parts(
+                        table.raw.contents() as *const f16,
+                        table.raw.length() as usize / 2,
+                    );
+                    for (i, &id) in ids.iter().enumerate() {
+                        let s = id as usize * dim;
+                        for j in 0..dim {
+                            o[i * dim + j] = t[s + j].to_f32();
+                        }
+                    }
+                }
             }
         }
     }
-
-    // ── Metal shader dispatches, no ctx flush ────────────────────────────
-    // These were CPU scalar loops on shared memory in the layer_forward_fused
-    // era (where the override bypassed the trait). Now that models call them
-    // directly they must encode into ctx.cmd() like every other GPU op so the
-    // single-command-buffer batching stays intact.
 
     fn split_qkv(
         ctx: &mut Self::Context,
@@ -223,6 +407,10 @@ impl Backend for MetalBackend {
         q_dim: usize,
         kv_dim: usize,
     ) {
+        let qkv = qkv.expect_f32("split_qkv qkv");
+        let q = q.expect_f32_mut("split_qkv q");
+        let k = k.expect_f32_mut("split_qkv k");
+        let v = v.expect_f32_mut("split_qkv v");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes
@@ -237,6 +425,8 @@ impl Backend for MetalBackend {
         tokens: usize,
         im: usize,
     ) {
+        let gu = gu.expect_f32("fused_silu_mul_split gate_up");
+        let out = out.expect_f32_mut("fused_silu_mul_split out");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.silu_mul_split_enc(enc, gu, out, tokens, im);
@@ -257,8 +447,11 @@ impl Backend for MetalBackend {
         eps: f32,
         mode: i32,
     ) {
-        // Fused norm (optional) + RoPE (optional) + transpose, all in one
-        // Metal dispatch. See ferrum_attention::metal::shaders::norm_rope.
+        let input = input.expect_f32("qk_norm_rope input");
+        let norm_w = norm_w.expect_f32("qk_norm_rope norm_w");
+        let cos = cos.expect_f32("qk_norm_rope cos");
+        let sin = sin.expect_f32("qk_norm_rope sin");
+        let output = output.expect_f32_mut("qk_norm_rope output");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.qk_norm_rope(
@@ -279,12 +472,11 @@ impl Backend for MetalBackend {
         nkv: usize,
         hd: usize,
     ) {
-        // Pre-allocated head-major append: data already laid out as
-        // [nkv, new_tokens, hd] (e.g. produced by qk_norm_rope). Writes into
-        // cache slot [nkv, cache_len .. cache_len+new_tokens, hd].
-        //
-        // No allocation; no transpose; one compute encoder on ctx.cmd().
         debug_assert!(cache_len + new_tokens <= cache_capacity);
+        let cache_k = cache_k.expect_f32_mut("kv_cache_append cache_k");
+        let cache_v = cache_v.expect_f32_mut("kv_cache_append cache_v");
+        let new_k_head_major = new_k_head_major.expect_f32("kv_cache_append new_k");
+        let new_v_head_major = new_v_head_major.expect_f32("kv_cache_append new_v");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.kv_cache_append(
@@ -318,8 +510,8 @@ impl Backend for MetalBackend {
         heads: usize,
         dim: usize,
     ) {
-        // [heads, tokens, dim] → [tokens, heads, dim] via dedicated Metal shader.
-        // Called after flash_attention to return Q in row-major layout for O-proj.
+        let src = src.expect_f32("transpose_head_to_token src");
+        let dst = dst.expect_f32_mut("transpose_head_to_token dst");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.transpose_out(enc, src, dst, tokens, heads, dim);
@@ -333,6 +525,8 @@ impl Backend for MetalBackend {
         rows: usize,
         cols: usize,
     ) {
+        let data = data.expect_f32_mut("add_bias data");
+        let bias = bias.expect_f32("add_bias bias");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.add_bias_enc(enc, data, bias, rows, cols);
@@ -349,6 +543,10 @@ impl Backend for MetalBackend {
         tokens: usize,
         dim: usize,
     ) {
+        let x = x.expect_f32("layer_norm x");
+        let gamma = gamma.expect_f32("layer_norm gamma");
+        let beta = beta.expect_f32("layer_norm beta");
+        let out = out.expect_f32_mut("layer_norm out");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes
@@ -357,6 +555,8 @@ impl Backend for MetalBackend {
     }
 
     fn gelu(ctx: &mut Self::Context, x: &Self::Buffer, out: &mut Self::Buffer, len: usize) {
+        let x = x.expect_f32("gelu x");
+        let out = out.expect_f32_mut("gelu out");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.gelu_enc(enc, x, out, len);
@@ -364,23 +564,98 @@ impl Backend for MetalBackend {
     }
 
     fn add_inplace(ctx: &mut Self::Context, r: &mut Self::Buffer, x: &Self::Buffer, len: usize) {
-        // Fused in-place add via add_enc with output == residual input.
-        // Metal's `add_f32` kernel reads a[tid] + b[tid] -> out[tid]; using the
-        // same buffer for a and out is well-defined because each thread handles
-        // exactly one element with no cross-thread dependency.
+        let r = r.expect_f32_mut("add_inplace r");
+        let x = x.expect_f32("add_inplace x");
         let cmd = ctx.cmd();
         let enc = cmd.new_compute_command_encoder();
         st().pipes.add_enc(enc, r, x, r, len);
         enc.end_encoding();
     }
 
+    // ── Buffer management ────────────────────────────────────────────
+
     fn alloc(len: usize) -> Self::Buffer {
-        st().pipes.buffer_empty(len)
+        // Scratch / output buffers are always f32 — the precision-sensitive
+        // compute paths expect it.
+        MetalBuf {
+            raw: alloc_f32_raw(len),
+            dtype: Dtype::F32,
+            n: len,
+        }
     }
+
     fn to_vec(buf: &Self::Buffer, len: usize) -> Vec<f32> {
-        MetalPipelines::read_buffer(buf, len)
+        match buf.dtype {
+            Dtype::F32 => {
+                let ptr = buf.raw.contents() as *const f32;
+                unsafe { std::slice::from_raw_parts(ptr, len).to_vec() }
+            }
+            Dtype::F16 => {
+                let ptr = buf.raw.contents() as *const f16;
+                let slice = unsafe { std::slice::from_raw_parts(ptr, len) };
+                slice.iter().map(|h| h.to_f32()).collect()
+            }
+        }
     }
+
     fn from_slice(data: &[f32]) -> Self::Buffer {
-        st().pipes.buffer_from_data(data)
+        // Activations, cos/sin tables, temporary scratch — all f32.
+        MetalBuf {
+            raw: buffer_from_f32_slice(data),
+            dtype: Dtype::F32,
+            n: data.len(),
+        }
+    }
+
+    fn from_weight_bytes(raw: &[u8], src_dtype: SrcDtype) -> Self::Buffer {
+        let n = raw.len() / src_dtype.bytes_per_elem();
+        let want_f16 = prefer_f16_weights() && n >= F16_MIN_ELEMS;
+
+        if !want_f16 {
+            // Default behaviour: materialise as f32, matches pre-refactor
+            // MetalBackend byte-for-byte.
+            let data = src_dtype.to_f32_vec(raw);
+            return MetalBuf {
+                raw: buffer_from_f32_slice(&data),
+                dtype: Dtype::F32,
+                n,
+            };
+        }
+
+        // f16 storage path — go directly from raw bytes where possible to
+        // avoid the transient 2× RAM spike.
+        match src_dtype {
+            SrcDtype::F16 => MetalBuf {
+                raw: buffer_from_f16_bytes(raw),
+                dtype: Dtype::F16,
+                n,
+            },
+            SrcDtype::BF16 => {
+                // bf16 → f16 via f32. Loses magnitude range (bf16 has a
+                // broader exponent) but gains mantissa precision, which for
+                // typical weight magnitudes (|w| < 32) is a net upgrade.
+                let mut f16_bytes = vec![0u8; n * 2];
+                for i in 0..n {
+                    let v = bf16::from_le_bytes([raw[i * 2], raw[i * 2 + 1]]).to_f32();
+                    let h = f16::from_f32(v).to_le_bytes();
+                    f16_bytes[i * 2] = h[0];
+                    f16_bytes[i * 2 + 1] = h[1];
+                }
+                MetalBuf {
+                    raw: buffer_from_f16_bytes(&f16_bytes),
+                    dtype: Dtype::F16,
+                    n,
+                }
+            }
+            SrcDtype::F32 => {
+                // f32 → f16 downcast. Halves storage.
+                let data = src_dtype.to_f32_vec(raw);
+                MetalBuf {
+                    raw: buffer_f16_from_f32(&data),
+                    dtype: Dtype::F16,
+                    n,
+                }
+            }
+        }
     }
 }
