+++ crates/ferrum-kernels/src/backend/mod.rs
@@ -13,8 +13,5 @@ pub mod cpu;
 #[cfg(feature = "metal")]
 pub mod metal;
 
-#[cfg(feature = "metal")]
-pub mod metal_f16;
-
 #[cfg(feature = "cuda")]
 pub mod cuda;
